requirements.txt
opencv-python
ultralytics
pyfirmata2
pyzmq
